Move S-bend TEDs to their own file
