+++ src/openrct2/libopenrct2.vcxproj
@@ -558,6 +558,7 @@
     <ClInclude Include="ride\rtd\water\SplashBoats.h" />
     <ClInclude Include="ride\rtd\water\SubmarineRide.h" />
     <ClInclude Include="ride\ted\PitchAndRoll.h" />
+    <ClInclude Include="ride\ted\TED.SBend.h" />
     <ClInclude Include="ride\ted\TrackElementDescriptor.h" />
     <ClInclude Include="ride\ted\TrackElemType.h" />
     <ClInclude Include="ride\ted\TrackGroup.h" />
